requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.9
Django==1.11.3
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
pycparser==2.18
pyOpenSSL==17.1.0
python-magic==0.4.13
pytz==2017.2
six==1.10.0
